feature: Parse streaming tokens as they come in to put code blocks into code cells
This requires some non-trivial parsing logic that can recognize when
the available buffered data is incomplete or ambiguous. E.g. if the tokens
we've recieved thus far are "\n``" then it may or may not be the case that
the next token will be something like "`wolfram\n", making the complete
buffer "\n```wolfram\n", which would mean we have to start a code cell.

* Move cell print logic to new CellPrint2 function in Utils.wl