Beginnings of getting persona (@) references to affect queries.
The queries successfully target the right persona, but also give a bunch of messages, mainly about uncaught Throws. My hope is that there's some simple bit of package convention that I'm failing to follow or something, but committing my progress as is in the hopes of getting more eyes on this.