hotfix: Fix pink boxes in output tab numbers in Chat-Driven notebooks
This commit fixes an issue with two pink boxes will tooltips saying
"$Aborted" being rendered in place of the current page number and
page count labels in tabbed ChatOutput cells.

I wasn't able to determine the exact cause of the bug, however some
things I learned:

* The pink boxes are shown when a ChatInput is evaluated a second time,
  causing reformatCell[..] to change the ChatOutput cell dingbat to
  "AssistantIconTabbed" template box.

* The pink boxes only appear in Chat-Driven notebooks, not in Chat-Enabled
  notebooks.

  - This seems to further becaused by the value of DefaultNewCellStyle.
    Doing:
      SetOptions[EvaluationNotebook[], DefaultNewCellStyle -> "Input"]
    reliably makes the bug go away. Changing "Input" to "ChatInput" reliably
    makes the bug return.

* Cmd+Shift+E to reveal the raw source of a "ChatOutput" cell and then
  toggling back to the rendered view fixes the pink boxes.

So the underlying cause of this bug doesn't appear to be a simple logic
error in the Chatbook code -- some more subtle FE behavior or bug is
interfering.

Consequently, this commit isn't a "direct" fix, more of a workaround;
it changes the page count rendering code to use different logic for Cloud
checking, so that the overall page count code can execute purely within
the FE's evaluator (no call to the Kernel). This appears to work around
the pink box bug, but I don't know why exactly.