Make sure the Notebook Assistant appears on the same monitor as the user's notebook.
Note that there is some risk to this change. This relies on NotebookTools`NotebookDisplayIndex, which was added to the kernel in 14.1, but which hasn't been widely tested on different platforms or monitor configurations. YMMV

Also note that there's a line of code, currently commented out, which will slide the user's notebook to the right if the Notebook Assistant would have obscured any part of it. That's commented out because it was not explicitly requested, but it does avoid a piece of the UX which is less than ideal IMO. Feel free to use it.