Modify scripts to be aware of dark mode support:
14.2 is aware of some features of dark mode, but not all. While Chatbook
supports 14.1 and 14.2, we build two sets of stylesheets. One set has no
dark mode settings, while the other set does. The DarkModeSupport components
are dark-mode aware and are built by running BuildStylesheets[All] from
a 14.3 session. If an older script runs BuildStylesheets then it won't
overwrite anything in the DarkModeSupport directory.

Notes on PacletManager extensions subtleties:

The PacletManager sorts extensions by their path. If we want the
FrontEnd extension for 14.3 to be found before the default, then it must
come alphabetically before the default. For that reason, we have

  {"FrontEnd", "Root" -> "FrontEnd", Prepend -> True}        (*original*)
  {"FrontEnd", "Root" -> "DarkModeSupport", Prepend -> True} (*new*)

where the original version contains the string resources, their
localizations, and non-dark-mode-aware stylesheets. The new version
contains dark-mode stylesheets and graphic resources.

They override the originals because their PrivatePaths are placed ahead
of the defaults in terms of search order. If a file isn't found in the
new locations, then the search continues to the original locations.